refactor(queries): extract trading-limits + oracle-snapshots + breaker config to lib/queries/config.ts
Follows PR #277 (lib/queries/pools.ts). Moves the 7 config-ish queries (ALL_TRADING_LIMITS, TRADING_LIMITS, ORACLE_SNAPSHOTS, ORACLE_SNAPSHOTS_CHART, ORACLE_SNAPSHOTS_COUNT_PAGE, POOL_DEPLOYMENT, POOL_BREAKER_CONFIG) verbatim out of lib/queries.ts and re-exports them through the existing barrel. The module-wide snapshot test in queries.test.ts pinned the move byte-for-byte; no test or import changes required.
