+++ ui-dashboard/src/lib/queries.ts
@@ -2,104 +2,7 @@
 // so this barrel can grow without breaking the per-domain organisation downstream.
 // Re-exported here to keep existing `from "@/lib/queries"` imports stable.
 export * from "./queries/pools";
-
-// Fetches all TradingLimit rows for a chain. With 2 rows per FPMM pool (one
-// per token), this stays well under Hasura's silent 1000-row cap for current
-// pool counts. If pool count exceeds ~500 per chain, add pagination.
-export const ALL_TRADING_LIMITS = `
-  query AllTradingLimits($chainId: Int!) {
-    TradingLimit(where: { chainId: { _eq: $chainId } }) {
-      id poolId token limit0 limit1 decimals
-      netflow0 netflow1
-      lastUpdated0 lastUpdated1
-      limitPressure0 limitPressure1
-      limitStatus updatedAtBlock updatedAtTimestamp
-    }
-  }
-`;
-
-export const TRADING_LIMITS = `
-  query TradingLimits($poolId: String!) {
-    TradingLimit(where: { poolId: { _eq: $poolId } }) {
-      id token limit0 limit1 decimals
-      netflow0 netflow1
-      lastUpdated0 lastUpdated1
-      limitPressure0 limitPressure1
-      limitStatus updatedAtBlock updatedAtTimestamp
-    }
-  }
-`;
-
-export const ORACLE_SNAPSHOTS = `
-  query OracleSnapshots($poolId: String!, $limit: Int!, $offset: Int!, $orderBy: [OracleSnapshot_order_by!]!) {
-    OracleSnapshot(
-      where: { poolId: { _eq: $poolId } }
-      order_by: $orderBy
-      limit: $limit
-      offset: $offset
-    ) {
-      id chainId
-      poolId
-      timestamp
-      oraclePrice
-      oracleOk
-      numReporters
-      priceDifference
-      rebalanceThreshold
-      source
-      blockNumber
-      txHash
-      deviationRatio
-      hasHealthData
-    }
-  }
-`;
-
-// Separate query for charts — always fetches the most recent N snapshots
-// ordered by timestamp desc, then reversed client-side for chronological display.
-// Decoupled from table pagination so charts always show full history context.
-export const ORACLE_SNAPSHOTS_CHART = `
-  query OracleSnapshotsChart($poolId: String!, $limit: Int!) {
-    OracleSnapshot(
-      where: { poolId: { _eq: $poolId } }
-      order_by: { timestamp: desc }
-      limit: $limit
-    ) {
-      id chainId
-      poolId
-      timestamp
-      oraclePrice
-      oracleOk
-      numReporters
-      priceDifference
-      rebalanceThreshold
-      source
-      blockNumber
-      txHash
-      deviationRatio
-      hasHealthData
-    }
-  }
-`;
-
-export const ORACLE_SNAPSHOTS_COUNT_PAGE = `
-  query OracleSnapshotsCountPage($poolId: String!, $limit: Int!, $offset: Int!) {
-    OracleSnapshot(where: { poolId: { _eq: $poolId } }, limit: $limit, offset: $offset) {
-      id
-    }
-  }
-`;
-
-export const POOL_DEPLOYMENT = `
-  query PoolDeployment($poolId: String!) {
-    FactoryDeployment(
-      where: { poolId: { _eq: $poolId } }
-      limit: 1
-    ) {
-      txHash
-    }
-  }
-`;
+export * from "./queries/config";
 
 // Preferred LP ownership query. Environments without LiquidityPosition should
 // show a migration message rather than attempt a correctness-risky fallback.
@@ -181,74 +84,6 @@ export const ALL_OLS_POOLS = `
       limit: 1000
     ) {
       poolId
-    }
-  }
-`;
-
-// Per-pool breaker config + recent trip history. Keyed by rateFeedID (NOT
-// pool ID) because a single feed underpins multiple pools (e.g.
-// USDC/USDm + axlUSDC/USDm both use feed 0xa1a8…), so denormalising onto
-// Pool would write-amplify on every breaker event. The dashboard fetches
-// this in parallel with POOL_DETAIL_WITH_HEALTH and renders <BreakerPanel />
-// only when the response is non-empty.
-//
-// Returns an array because, in principle, a feed could have multiple
-// trip-able breakers; in production today every v3 feed has exactly one
-// MedianDelta or ValueDelta plus the always-present MarketHours row (the
-// latter has no per-feed config and is filtered client-side via
-// `breaker.kind === "MARKET_HOURS"` to drive the title-row pill).
-export const POOL_BREAKER_CONFIG = `
-  query PoolBreakerConfig($chainId: Int!, $rateFeedID: String!) {
-    BreakerConfig(
-      where: {
-        chainId: { _eq: $chainId }
-        rateFeedID: { _eq: $rateFeedID }
-      }
-    ) {
-      id
-      enabled
-      cooldownTime
-      rateChangeThreshold
-      smoothingFactor
-      medianRatesEMA
-      referenceValue
-      lastMedianRate
-      lastUpdatedAt
-      status
-      tradingMode
-      lastStatusUpdatedAt
-      cooldownEndsAt
-      lastTripAt
-      lastTripTxHash
-      lastResetAt
-      tripCountLifetime
-      breaker {
-        id
-        address
-        kind
-        activatesTradingMode
-        defaultCooldownTime
-        defaultRateChangeThreshold
-      }
-    }
-    BreakerTripEvent(
-      where: {
-        chainId: { _eq: $chainId }
-        rateFeedID: { _eq: $rateFeedID }
-      }
-      order_by: [{ blockTimestamp: desc }]
-      limit: 50
-    ) {
-      id
-      blockTimestamp
-      txHash
-      medianRateAtTrip
-      referenceAtTrip
-      thresholdAtTrip
-      breaker {
-        address
-        kind
-      }
     }
   }
 `;
